+++ source/core/database/src/main/kotlin/com/xayah/core/database/DatabaseModule.kt
@@ -26,7 +26,8 @@ object DatabaseModule {
             .enableMultiInstanceInvalidation()
             .addMigrations(
                 DatabaseMigrations.MIGRATION_7_8,
-                DatabaseMigrations.MIGRATION_8_9  // 添加这个
+                DatabaseMigrations.MIGRATION_8_9,
+                DatabaseMigrations.MIGRATION_9_10  // 添加这个
             )
             .build()