+++ source/core/database/src/main/kotlin/com/xayah/core/database/dao/MediaDao.kt
@@ -18,6 +18,12 @@ interface MediaDao {
     @Upsert(entity = MediaEntity::class)
     suspend fun upsert(item: MediaEntity)
 
+    @Query("UPDATE MediaEntity SET extraInfo_isCanceled = 1 WHERE indexInfo_backupTimestamp = :timestamp")
+    suspend fun markAsCanceledByTimestamp(timestamp: Long)
+
+    @Query("DELETE FROM MediaEntity WHERE indexInfo_backupTimestamp = :timestamp AND extraInfo_isCanceled = 1")
+    suspend fun deleteCanceledByTimestamp(timestamp: Long)
+
     @Query(
         "SELECT * FROM MediaEntity" +
                 " WHERE indexInfo_opType = :opType AND indexInfo_name = :name" +